+++ etc/blocks/pool-explorer/block/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @milaboratories/milaboratories.pool-explorer
 
+## 1.0.114
+
+### Patch Changes
+
+- Updated dependencies [8c3b22b]
+  - @platforma-sdk/model@1.30.11
+  - @milaboratories/milaboratories.pool-explorer.ui@1.1.77
+  - @milaboratories/milaboratories.pool-explorer.model@1.0.36
+
+## 1.0.113
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.ui@1.1.76
+
 ## 1.0.112
 
 ### Patch Changes