+++ etc/blocks/ui-examples/ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @milaboratories/milaboratories.ui-examples.ui
 
+## 1.3.75
+
+### Patch Changes
+
+- Updated dependencies [8c3b22b]
+  - @platforma-sdk/model@1.30.11
+  - @milaboratories/milaboratories.ui-examples.model@1.1.54
+
+## 1.3.74
+
+### Patch Changes
+
+- 764ea86: useAgGridOptions: add "loadingText" option, fully ignore "loadingOverlayComponentParams" completely (show warning if used)
+
 ## 1.3.73
 
 ### Patch Changes